requirements.txt
django
django-allauth
requests-cache
python-memcached